api/requirements.txt
flask
Flask-SocketIO==5.3.1
Flask-Smorest
Flask-SQLAlchemy
Flask_Cors
Flask-Migrate
flask_jwt_extended
passlib
marshmallow
psycopg2
datetime
